requirements.txt
anyio==4.8.0
argcomplete==3.6.3
asgiref==3.10.0
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.14.2
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.72
botocore==1.40.72
Brlapi==0.8.6
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-350-build/cockpit-350/tmp/wheel/cockpit-350-py3-none-any.whl#sha256=e9cb30e75d498ef4db7f9bc5812d340005d173418bb52a80db76b1124bbfcf1d
contourpy==1.3.3
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
distro==1.9.0
Django==5.2.7
django-formtools==2.5.1
django-otp==1.6.3
django-phonenumber-field==8.4.0
django-two-factor-auth==1.18.1
dnf==4.24.0
fedora-third-party==0.10
file-magic==0.4.0
fonttools==4.60.1
fros==1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.12.0
idna==3.10
iso639==0.1.4
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.22
libdnf==0.75.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
matplotlib==3.10.6
nftables==0.1
numpy==2.3.3
olefile==0.47
packaging==24.2
pandas==2.3.3
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
ply==3.11
productmd==1.45
progressbar2==4.5.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pyOpenSSL==25.0.0
pyparsing==3.2.5
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.7-build/kernel-6.17.7/linux-6.17.7-200.fc42.x86_64/tools/net/ynl
qrcode==8.2
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
sqlparse==0.5.3
systemd-python==235
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xkbregistry==0.3
